abm {hte}	R Documentation

Agent-based colonization/decolonization model

Description

Given patient movement data, simulated transmission dynamics and screening.

Usage

run_abm(
  data,
  b = 0.03,
  gamma = 0.02,
  lambda0 = 0,
  p0 = 0.05,
  alpha = 0.01,
  beta = 0.15,
  adm_eff = 0,
  status_changes = FALSE,
  outside = "out",
  verbose = getOption("verbose", TRUE)
)

Arguments

data
a data frame containing patient movement and testing information. It should have at least the following columns:

patient

a unique identifier for each patient

unit

a unique identifier for each unit

time

a number representing time

event

a character vector specifying the kind of event: admission, discharge, transfer, or test.

b
transmission rate

gamma
decolonization rate

lambda0
background colonization rate

p0
probability of colonization on admission

alpha, beta
false positive and negative testing error rates

adm_eff
numeric; efficiency of testing on admission

status_changes
logical; should true times of colonization or decolonization events be included in the output?

outside
character; name of "outside" unit

verbose
run-time information?

Details

run_abm simulated the spread of a transmissible agent through a hospital. When testing events are given in data, all patients on the given unit are tested. Patients are screened on admission.

Value

run_abm returns a tibble with one row per testing event. This contains information on the test result, true colonization status, and location of the tested patient.

See Also

More on simulated data: fake_data, simuldat()

More on the transmission model: transmission

Examples

## Not run: 

  library(tidyverse)

  expand_grid(
    time=seq(70,1827,by=7),
    unit=LETTERS[1:9],
    event="test"
  ) -> testsched

  sim_pat_mov |>
    bind_rows(testsched) |>
    arrange(time) |>
    run_abm(verbose=TRUE) -> dat

## End(Not run)

[Package hte version 0.0.19.3 Index]