Bernoulli_filter {hte}	R Documentation

Bernoulli filter

Description

Filter the individual patient data with given forces of infection, recovery rates, and test characteristics.

Usage

Bernoulli_filter(data, lambda, gamma, theta)

Bfilter(data, theta)

Arguments

data
data set

lambda
unit-specific force of infection

gamma
unit-specific recovery rate

theta
list of parameters

Details

Bernoulli_filter runs a Bernoulli filter, updating the expected prevalence. It returns the log likelihood, occupancy, and expected prevalences.

Bfilter uses an alternative algorithm.

Value

Bernoulli_filter returns a tibble containing the expected prevalences (for both isolated and un-isolated patients), unit occupancy, force of infection, and conditional log likelihood for each unit at each event time. The sum of the log likelihood column (logLik) is the log likelihood of the data.

Bfilter returns the log likelihood of the furnished data.

Examples

library(tidyverse)

set.seed(626292345)

fake_data |>
  arrange(patient,time) -> dat

theta <- list(
  lambda=c(out=0,A=0.01,B=0.1,C=0.2,D=0.3,E=0.5),
  gamma=c(out=0.1,A=0.1,B=0.1,C=0.1,D=0.1,E=0.1),
  p0=0.2,
  isol_factor=0.1,
  alpha=0.05,
  beta=0.2
)

dat |>
  Bfilter(theta) -> ll1
sum(ll1)

dat |>
  Bernoulli_filter(
    lambda=theta$lambda,
    gamma=theta$gamma,
    theta
  ) -> f
f |> filter(logLik!=0) |> pull(logLik) -> ll2
sum(ll2)

f

f |>
  select(unit,time,prev_i,prev_u) |>
  pivot_longer(c(prev_i,prev_u)) |>
  group_by(unit) |>
  ## prevalence is not estimated outside the hospital
  filter(!all(is.na(value))) |>
  ungroup() |>
  ggplot(aes(x=time,color=name,y=value))+
  geom_line(alpha=0.5)+
  scale_color_manual(values=c(prev_i="blue",prev_u="red"))+
  facet_grid(unit~.,labeller=label_both)+
  labs(y="prevalence")+
  theme_bw()

[Package hte version 0.0.11.0 Index]