independent {hte}	R Documentation

Independent infection model

Description

Model under which each patient's trajectory of infection is independent, conditional on the force of infection.

Usage

indep_homog_filter(params, data)

indep_homog_objfun(params, data, est = character(0))

indep_unit_spec_filter(params, data)

indep_unit_spec_objfun(params, data, est = character(0))

Arguments

params
named vector of parameters

data
patient movement, isolation, and testing data

est
names of parameters to estimate

Details

Parameters in the model include:

lambda

force of infection. In the homogeneous model, this is constant across units in the hospital; in the unit-specific model, there is one value of lambda for each unit.

lambda.out

force of infection outside the hospital

gamma

recovery rate

alpha,beta

false positive and negative probabilities

p0

probability of infection on admission

isol_factor

multiplicative effect of contact isolation on susceptibility

eta

additional information on infection status

indep_homog_filter runs a Bernoulli filter for the independent model with a global lambda and gamma.

indep_homog_objfun is a stateful objective function for the independent model with a global lambda and gamma.

indep_unit_spec_filter runs a Bernoulli filter for the independent model with a unit-specific lambda.

indep_unit_spec_objfun is a stateful objective function for the independent model with unit-specific lambda.

See Also

More on the independent infection model: simuldat()

More on stateful objective functions: stobfun(), transmission

Examples

library(tidyverse)
library(lubridate)
library(pomp)
library(hte)

set.seed(626292345)

## Select some data:
fake_data |>
  filter(
    event!="test" | time < 250,
    time < 1000
  ) |>
  arrange(patient,time) -> dat

## Create an objective function:
indep_homog_objfun(
  params=c(
    lambda.out=0.05,lambda=0.1,gamma=0.01,p0=0.1,
    isol_factor=0.2,alpha=0.02,beta=0.1,eta = 0.5
  ),
  est=c("lambda","gamma"),
  data=dat
) -> f

## Fit the model:
optim(
  par=log(c(0.1,0.01)),
  fn=f,
  control=list(reltol=1e-3)
) -> out
f(out$par)
coef(f)

## Examine the filter results:
indep_homog_filter(params=coef(f),data=dat) -> ff
ff |> filter(logLik!=0)

## Construct a slice:
indep_homog_objfun(
  params=coef(f),
  est=c("lambda"),
  data=dat
) -> g

log.lambda <- seq(log(0.001),to=log(1),length=10)
plot(exp(log.lambda),sapply(log.lambda,g),log='x')

## Construct an objective function for the unit-specific model:
theta <- coef(f)
indep_unit_spec_objfun(
  params=c(
    lambda=setNames(rep.int(theta["lambda"],6),unique(dat$unit)),
    theta[c("p0","gamma","alpha","beta","isol_factor","eta")]
  ),
  est=c(
    "lambda.A","lambda.B","lambda.C","lambda.D","lambda.E","lambda.out","gamma"
  ),
  data=dat
) -> h

## Fit the model:
optim(
  par=log(coef(h)[
    c("lambda.A","lambda.B","lambda.C","lambda.D","lambda.E","lambda.out","gamma")
  ]),
  fn=h,
  control=list(reltol=1e-3)
) -> out
h(out$par)
coef(h)

[Package hte version 0.0.19.1 Index]