simuldat {hte}	R Documentation

simuldat

Description

simuldat simulates data representing the flow of a body of patients through a hospital over a specified window of time.

simul_patient simulates a single patient's history of movement, testing, isolation, and infection.

Usage

simuldat(
  nbeds = c(50, 60),
  arrival = 20,
  window = c("1999-12-31T23:59:59+0000", "2003-01-01T00:00:00+0000"),
  units = list(A = list(shape = 10, scale = 2/10), B = list(shape = 5, scale = 0.5/5), C
    = list(shape = 1, scale = 0.5/1), D = list(shape = 1, scale = 3/1), E = list(shape =
    0.2, scale = 8/0.2), out = list(shape = 0.5, scale = 300/0.5)),
  visits = list(size = 0.5, mu = 9),
  uperv = list(size = 1, mu = 0.5),
  min_dur = 1/24,
  testing_freq = c(A = NA, B = NA, C = 1/7, D = 1/7, E = 1/7, out = NA),
  isolation = list(on = 1/50, off = 1/50),
  infection = list(lambda = c(A = 0.01, B = 0.02, C = 0.001, D = 0.5, E = 0.1, out =
    0.05), gamma = 0.01, p0 = 0.1, isol_factor = 0.2),
  alpha = 0.02,
  beta = 0.1,
  verbose = getOption("verbose", TRUE)
)

simul_patient(
  patient,
  t0,
  tf,
  units,
  visits,
  uperv,
  min_dur,
  testing_freq,
  isolation,
  infection,
  alpha,
  beta
)

indep_infect(lambda, gamma, p0, isol_factor, times, loc, isol)

Arguments

nbeds
upper and lower bounds on number of beds

arrival
Poisson arrival rate of new patients

window
window of simulation

units
a named list with one entry per unit. Each entry is itself a list with the parameters of the Gamma-distribution for the duration of stay in the unit.

visits
list containing parameters of a negative binomial distribution for the number of visits per patient.

uperv
list containing parameters for a negative binomial distribution for the number of units visited per visit.

min_dur
minimum duration of stay in any unit

testing_freq
named numeric vector of unit-specific testing frequencies.

isolation
list containing parameters of the isolation model

infection
parameters of the infection model

alpha, beta
false positive and negative testing error rates

verbose
run-time information?

patient
patient name or number

t0, tf
initial and final times of patient itinerary

lambda
force of infection

gamma
recovery rate

p0
initial probability of infection

isol_factor
reduction in susceptibility due to isolation

times
times at which status is reported

loc
location of patient at each time

isol
isolation status

Details

Simulate hospital movement, testing, and isolation data.

Value

infection status vector

See Also

More on simulated data: fake_data

More on the independent infection model: independent

Examples

library(tidyverse)
library(lubridate)

## Examine the data:
fake_data

## Verify certain conditions hold:
stopifnot(
  `admission condition violation`=fake_data |>
    group_by(patient,visit) |> slice_head() |>
    filter(event!="admit") |> nrow()==0,
  `discharge condition violation`=fake_data |>
    group_by(patient,visit) |> slice_tail() |>
    filter(event!="discharge",event!="stop") |> nrow()==0,
  `unit violation`=fake_data |> filter(is.na(unit)) |> nrow()==0,
  `event violation`=fake_data |> filter(is.na(event)) |> nrow()==0
)

fake_data |>
  mutate(
    time=as.numeric(
      as.duration(
        interval(date,start="2000-01-01T00:00:00+0000")
      ),
      units="day"
    )
  ) -> fake_data

fake_data |>
  group_by(patient,visit) |>
  summarize(dur=max(time)-min(time)) |>
  ungroup() |>
  group_by(patient) |>
  summarize(dur=sum(dur)) |>
  ggplot(aes(x=log10(dur)))+
  geom_histogram(bins=20)+
  labs(title="total duration of hospitalization")+
  theme_bw()

fake_data |>
  group_by(patient,visit) |>
  summarize(dur=max(time)-min(time)) |>
  ungroup() |>
  ggplot(aes(x=log10(dur)))+
  geom_histogram(bins=40)+
  labs(title="duration of hospital visit")+
  theme_bw()

fake_data |>
  filter(
    event!="test",
    event!="isolate",
    event!="release",
    event!="stop"
  ) |>
  group_by(patient,visit) |>
  arrange(time) |>
  mutate(dur=lead(time)-time) |>
  ungroup() |>
  filter(unit!="out",!is.na(dur)) |>
  ggplot(aes(x=log10(dur),fill=unit,group=unit))+
  geom_histogram(aes(y=after_stat(density)),bins=40)+
  facet_grid(unit~.,scales="free_y")+
  labs(title="duration of stay by unit")+
  theme_bw()

fake_data |>
  group_by(patient) |>
  summarize(ntest=sum(event=="test")) |>
  ungroup() |>
  ggplot(aes(x=ntest))+
  geom_histogram(binwidth=1,center=0)+
  labs(title="number of tests per patient")+
  theme_bw()

fake_data |>
  arrange(time) |>
  mutate(
    dn=case_when(
      event=="admit"~1L,
      event=="discharge"~-1L,
      TRUE~0L
    ),
    occ=cumsum(dn)
  ) |>
  ggplot(aes(x=date,y=occ))+
  geom_step()+
  labs(title="hospital occupancy")+
  theme_bw()

fake_data |>
  arrange(time) |>
  select(date,test.result=result,isol,infected) |>
  pivot_longer(c(test.result,isol,infected)) |>
  filter(!is.na(value)) |>
  ggplot(aes(x=date,y=value,color=name))+
  geom_point()+
  geom_smooth()+
  guides(color="none")+
  labs(
    title="infection and isolation status, test results",
    y=""
  )+
  facet_grid(name~.)+
  theme_bw()

fake_data |>
  filter(event=="test") |>
  mutate(
    interval=cut(time,breaks=72,ordered=TRUE)
  ) |>
  select(interval,time,infected,isol,result) |>
  pivot_longer(c(infected,isol,result)) |>
  group_by(name,interval) |>
  summarize(
    time=mean(time),
    prev=mean(value),
    n=n()
  ) |>
  ungroup() |>
  ggplot(aes(x=time,y=prev,group=name,fill=name))+
  geom_col(position="dodge")+
  labs(title="infection, isolation, and detection through time")+
  theme_bw()+
  theme(axis.text.x=element_text(angle=90))

fake_data |>
  filter(event=="test") |>
  select(infected,result) |>
  count(infected,result) |>
  group_by(infected) |>
  mutate(prob=n/sum(n)) |>
  ungroup()

[Package hte version 0.0.12.0 Index]