stobfun {hte}	R Documentation

Stateful objective functions

Description

Convenience functions for constructing and working with stateful objective functions (‘stobfun’-class objects).

Usage

stobfun(
  embed,
  params,
  est = character(0),
  log = character(0),
  logit = character(0),
  objfun,
  data,
  ...
)

filterfun(embed, params, filtfun, data, ...)

transf_fns(log = character(0), logit = character(0), est = character(0))

embedding(...)

## S3 method for class 'stobfun'
coef(object, ...)

Arguments

embed
embedding (see embedding).

params
vector of parameters

est
character: names of parameters to be estimated

log
character; names of parameters to log transform.

logit
character; names of parameters to logit transform.

objfun
underlying objective function

data
data

...
When furnished to stobfun, additional arguments are passed to objfun. When furnished to embedding, arguments define th embedding. When furnished to coef, additional arguments are ignored.

filtfun
the function that actually applies the filter

object
‘stobfun’-class stateful objective function

Details

objfun will be called as objfun(theta,data), where theta is the nested list constructed according to the given specifications and data is the data.

Value

transf_fns returns a list of two functions. The first is the transformation to the estimation scale; the second is its inverse.

embedding returns the embedding function corresponding to the given specification.

coef(f) returns the parameter vector corresponding to the last call of the stateful objective function f.

Construction and usage of stateful objective functions

A stateful objective function is an ordinary function that can be used as an objective function in an optimization problem. In particular, it can be passed to optimizers such as optim, subplex, or nloptr. It is stateful in the sense that it remembers the argument with which it was last called.

To construct a stateful objective function, call the constructor function for the model of interest. The constructor function requires that you pass a vector of model parameters: this gives the default parameter values. It also requires that you pass the data and the names of the parameters that you wish to estimate. The constructor will return an object of class ‘stobfun’.

Having constructed a ‘stobfun’ stateful objective function, you can pass this to any suitable optimizer. Once the optimizer has returned, it is important that you call the function one last time, at the parameters the optimizer has returned (see examples). This ensures that the stored parameters are those at the (putative) optimum. You can retrieve these parameters via a call to coef.

See Also

More on stateful objective functions: independent, transmission

[Package hte version 0.0.16.0 Index]