independent {hte}	R Documentation

Independent infection model

Description

Model under which each patient's trajectory of infection is independent, conditional on the force of infection.

Usage

indep_homog_objfun(params, data, est = character(0))

indep_unit_spec_objfun(params, data, est = character(0))

Arguments

params
named vector of parameters

data
patient movement, isolation, and testing data

est
names of parameters to estimate

Details

indep_homog_objfun is a stateful objective function for the independent model with a global lambda and gamma.

indep_unit_spec_objfun is a stateful objective function for the independent model with unit specific lambda, one gamma for inside the hospital, and one gamma outside the hospital.

See Also

More on the independent infection model: simuldat()

More on stateful objective functions: stobfun(), transmission

Examples

library(tidyverse)
library(lubridate)
library(pomp)
library(hte)

set.seed(626292345)

fake_data |>
  ## filter out tests prior to day 250
  filter(
    event!="test" | time < 250,
    time < 1000
  ) |>
  arrange(patient,time) -> dat

indep_homog_objfun(
  params=c(
    lambda=0.1,gamma=0.01,p0=0.1,
    isol_factor=0.2,alpha=0.02,beta=0.1
  ),
  est=c("lambda","gamma"),
  data=dat
) -> f

optim(
  par=log(c(0.1,0.01)),
  fn=f,
  control=list(reltol=1e-3)
) -> out
f(out$par)
coef(f)

indep_homog_objfun(
  params=coef(f),
  est=c("lambda"),
  data=dat
) -> g

log.lambda <- seq(log(0.001),to=log(1),length=10)
plot(exp(log.lambda),sapply(log.lambda,g),log='x')

theta <- coef(f)
indep_unit_spec_objfun(
  params=c(
    lambda=setNames(rep.int(theta["lambda"],6),unique(dat$unit)),
    gamma.out=unname(theta["gamma"]),
    theta[c("p0","gamma","alpha","beta","isol_factor")]
  ),
  est=c(
    "lambda.A","lambda.B","lambda.C","lambda.D","lambda.E","lambda.out",
    "gamma","gamma.out"
  ),
  data=dat
) -> h

## Not run: 

optim(
  par=log(coef(h)[
    c("lambda.A","lambda.B","lambda.C","lambda.D","lambda.E","lambda.out",
      "gamma","gamma.out")
  ]),
  fn=h,
  control=list(reltol=1e-3)
) -> out
h(out$par)
coef(h)

## End(Not run)

[Package hte version 0.0.6.1 Index]