transmission {hte}	R Documentation

Transmission model

Description

Model under which the force of infection in each unit is proportional to the prevalence of infection in that unit.

Usage

trans_homog_filter(params, data, tol = 1e-04, maxit = 10)

trans_homog_objfun(params, data, est = character(0), tol = 1e-04, maxit = 10)

trans_unit_spec_filter(params, data, tol = 1e-04, maxit = 10)

trans_unit_spec_objfun(
  params,
  data,
  est = character(0),
  tol = 1e-04,
  maxit = 10
)

Arguments

params
named vector of parameters

data
patient movement, isolation, and testing data

tol
positive scalar; convergence tolerance (mean difference).

maxit
scalar integer; maximum number of fixed-point iterations. If tol is not achieved in maxit or fewer iterations, an error is generated.

est
names of parameters to estimate

Details

The basic transmission model assumes that the force of infection on an unisolated patient is

\lambda=b(P_u+aP_i)+\iota,

where P_u, P_i are the prevalences among unisolated and isolated patients, respectively and \iota represents the risk of acquisition unrelated to local prevalence. On an isolated patient, this force of infection is reduced by the factor isol_factor.

Parameters in the model include:

b

transmission rate. In the homogeneous model, this is constant across units in the hospital; in the unit-specific model, there is one value of b for each unit.

lambda.out

force of colonization outside the hospital

iota

baseline force of colonization

gamma

recovery rate

alpha,beta

false positive and negative probabilities

p0

probability of infection on admission

a

multiplicative effect of contact isolation on transmissibility

isol_factor

multiplicative effect of contact isolation on susceptibility

trans_homog_filter runs a fixed-point Bernoulli filter for the transmission model with global b and gamma.

trans_homog_objfun is a stateful objective function for the transmission model with global b and gamma.

trans_unit_spec_filter runs a fixed-point Bernoulli filter for the transmission model with unit-specific b.

trans_unit_spec_objfun is a stateful objective function for the transmission model with unit-specific transmission rates, an out-of-hospital force of infection parameter, and recovery rates that can be different inside and outside of hospital.

See Also

More on stateful objective functions: independent, stobfun()

Examples

library(tidyverse)
library(hte)

set.seed(339613584)

fake_data |>
  ## filter out tests prior to day 250
  filter(
    event!="test" | time < 250,
    time < 1000
  ) |>
  select(-infected) |>
  arrange(patient,time) -> dat

trans_homog_objfun(
  params=c(
    a=1,b=0.1,gamma=0.01,p0=0.1,lambda.out=0.01,
    iota=0,isol_factor=0.2,alpha=0.02,beta=0.1
  ),
  est=c("a","b"),
  data=dat
) -> f
f(log(c(0.05,0.4)))
coef(f)

optim(
  par=log(c(0.5,0.4)),
  fn=f,
  control=list(reltol=1e-2)
) -> out
f(out$par)
coef(f)

trans_homog_filter(params=coef(f),data=dat) -> ff
ff |> filter(logLik!=0)

trans_unit_spec_objfun(
  params=c(
    a=1,
    b.A=0.1,b.B=0.1,b.C=0.1,b.D=0.1,b.E=0.1,
    gamma=0.01,p0=0.1,
    lambda.out=0.1,
    iota=0,
    isol_factor=0.2,
    alpha=0.02,beta=0.1
  ),
  data=dat
) -> f
f()
coef(f)

[Package hte version 0.0.14.0 Index]